added prompts (cereal, guessing game, russian restaurant)

+++ ai_conversations/text_chatbot.py
@@ -33,8 +33,9 @@ class TextChatbot:
         self.response_number = 0
 
         self.OUTPUT_FOLDER = os.path.join(os.getcwd(), "outputs")
-        import shutil
-        shutil.rmtree(self.OUTPUT_FOLDER)
+        if os.path.exists(self.OUTPUT_FOLDER):
+            import shutil
+            shutil.rmtree(self.OUTPUT_FOLDER)
         os.makedirs(self.OUTPUT_FOLDER)
 
 
@@ -48,8 +49,8 @@ class TextChatbot:
     def get_chat_gpt_response_threaded(self, input_text):
         self.chat_history.append({"role": "user", "content": input_text})
         response = client.chat.completions.create(
-            #model="gpt-3.5-turbo", messages=self.chat_history
-            model="gpt-4", messages=self.chat_history
+            model="gpt-3.5-turbo", messages=self.chat_history
+            #model="gpt-4", messages=self.chat_history
         )
         self.chat_gpt_response = response.choices[0].message.content
         self.chat_history.append({"role": "assistant", "content": self.chat_gpt_response})
@@ -81,7 +82,7 @@ class TextChatbot:
                 self.response_number = self.response_number + 1
 
             except FileNotFoundError:
-                time.sleep(2)
+                time.sleep(5)
                 continue
 
             print("input text: " + input_text)
